Some better cleanup
